+++ packages/app/src/specs/banners/LoginBanner.vue
@@ -8,6 +8,12 @@
     class="mb-16px"
     :icon="ConnectIcon"
     dismissible
+    :has-banner-been-shown="hasBannerBeenShown"
+    :event-data="{
+      campaign: 'Log In',
+      medium: 'Specs Login Banner',
+      cohort: '' // TODO Connect cohort
+    }"
     @update:model-value="value => emit('update:modelValue', value)"
   >
     <p class="mb-24px">
@@ -50,8 +56,10 @@ query LoginBanner {
 
 withDefaults(defineProps<{
   modelValue: boolean
+  hasBannerBeenShown: boolean
 }>(), {
   modelValue: false,
+  hasBannerBeenShown: true,
 })
 
 const emit = defineEmits<{