+++ packages/app/src/specs/banners/RecordBanner.vue
@@ -8,6 +8,12 @@
     class="mb-16px"
     :icon="RecordIcon"
     dismissible
+    :has-banner-been-shown="hasBannerBeenShown"
+    :event-data="{
+      campaign: 'Record Runs',
+      medium: 'Specs Record Runs Banner',
+      cohort: '' // TODO Connect cohort
+    }"
     @update:model-value="value => emit('update:modelValue', value)"
   >
     <p class="mb-24px">
@@ -54,8 +60,10 @@ query RecordBanner {
 
 withDefaults(defineProps<{
   modelValue: boolean
+  hasBannerBeenShown: boolean
 }>(), {
   modelValue: false,
+  hasBannerBeenShown: true,
 })
 
 const emit = defineEmits<{