+++ packages/app/vite.config.mjs
@@ -2,6 +2,7 @@ import { makeConfig } from '../frontend-shared/vite.config.mjs'
 import Layouts from 'vite-plugin-vue-layouts'
 import Pages from 'vite-plugin-pages'
 import Copy from 'rollup-plugin-copy'
+import Legacy from '@vitejs/plugin-legacy'
 import { resolve } from 'path'
 
 export default makeConfig({
@@ -29,5 +30,10 @@ export default makeConfig({
         dest: 'dist',
       }],
     }),
+    Legacy({
+      targets: ['Chrome >= 64', 'Firefox >= 86', 'Edge >= 79'],
+      modernPolyfills: true,
+      renderLegacyChunks: false,
+    })
   ],
 })