+++ packages/driver/src/cy/commands/sessions/index.ts
@@ -53,6 +53,10 @@ export default function (Commands, Cypress, cy) {
 
   Commands.addAll({
     session (id, setup?: Function, options: { validate?: Function } = {}) {
+      if (Cypress.isBrowser('webkit')) {
+        return $errUtils.throwErrByPath('webkit.session')
+      }
+
       throwIfNoSessionSupport()
 
       if (!id || !_.isString(id) && !_.isObject(id)) {