+++ system-tests/test/web_security_spec.js
@@ -72,6 +72,7 @@ describe('e2e web security', () => {
 
   context('when enabled', () => {
     systemTests.it('fails', {
+      browser: '!webkit', // TODO(webkit): fix+unskip
       spec: 'web_security.cy.js',
       config: {
         experimentalSessionAndOrigin: true,
@@ -98,12 +99,13 @@ describe('e2e web security', () => {
     systemTests.it('displays warning when firefox and chromeWebSecurity:false', {
       spec: 'simple_passing.cy.js',
       snapshot: true,
+      // TODO(webkit): run this test in webkit
       browser: 'firefox',
       config: {
         chromeWebSecurity: false,
       },
       onStdout (stdout) {
-        expect(stdout).include('Your project has set the configuration option: chromeWebSecurity to false\n\nThis option will not have an effect in Firefox.')
+        expect(stdout).include('Your project has set the configuration option: `chromeWebSecurity` to `false`.\n\nThis option will not have an effect in Firefox.')
       },
     })
   })